+++ public/js/themes.js
@@ -1,0 +1,20 @@
+
+module.exports = {
+    "bootstrap-default": "/css/bootstrap-default.css",
+    "bootstrap-cerulean": "/css/bootstrap-cerulean.css",
+    "bootstrap-cosmo": "/css/bootstrap-cosmo.css",
+    "bootstrap-cyborg": "/css/bootstrap-cyborg.css",
+    "bootstrap-darkly": "/css/bootstrap-darkly.css",
+    "bootstrap-flatly": "/css/bootstrap-flatly.css",
+    "bootstrap-journal": "/css/bootstrap-journal.css",
+    "bootstrap-lumen": "/css/bootstrap-lumen.css",
+    "bootstrap-paper": "/css/bootstrap-paper.css",
+    "bootstrap-readable": "/css/bootstrap-readable.css",
+    "bootstrap-sandstone": "/css/bootstrap-sandstone.css",
+    "bootstrap-simplex": "/css/bootstrap-simplex.css",
+    "bootstrap-slate": "/css/bootstrap-slate.css",
+    "bootstrap-spacelab": "/css/bootstrap-spacelab.css",
+    "bootstrap-superhero": "/css/bootstrap-superhero.css",
+    "bootstrap-united": "/css/bootstrap-united.css",
+    "bootstrap-yeti": "/css/bootstrap-yeti.css"
+}